+++ build.sh
@@ -285,21 +285,43 @@ function build {
             sign_win "$destination"
         fi
     done
+
+    if [[ "$current_target" == "aarch64-pc-windows-msvc" ]]; then
+        # TODO: We still ship x64 OpenVPN with ARM64, so we need an x64 talpid-openvpn-plugin
+        # to include in the package.
+        log_info "Workaround: building x64 talpid-openvpn-plugin"
+        cargo build --target x86_64-pc-windows-msvc "${CARGO_ARGS[@]}" -p talpid-openvpn-plugin --lib
+        cp "$CARGO_TARGET_DIR/x86_64-pc-windows-msvc/$RUST_BUILD_MODE/talpid_openvpn_plugin.dll" "dist-assets/aarch64-pc-windows-msvc/talpid_openvpn_plugin.dll"
+        if [[ "$SIGN" == "true" ]]; then
+            sign_win "dist-assets/talpid_openvpn_plugin.dll"
+        fi
+    fi
 }
 
 if [[ "$(uname -s)" == "MINGW"* ]]; then
-    log_header "Building C++ code in $CPP_BUILD_MODE mode"
-    CPP_BUILD_MODES=$CPP_BUILD_MODE IS_RELEASE=$IS_RELEASE ./build-windows-modules.sh
+    for t in "${TARGETS[@]:-"x86_64-pc-windows-msvc"}"; do
+        case $t in
+            x86_64-pc-windows-msvc) CPP_BUILD_TARGET=x64;;
+            aarch64-pc-windows-msvc) CPP_BUILD_TARGET=ARM64;;
+            *)
+                log_error "Unknown Windows target: $t"
+                exit 1
+                ;;
+        esac
 
-    if [[ "$SIGN" == "true" ]]; then
-        CPP_BINARIES=(
-            "windows/winfw/bin/x64-$CPP_BUILD_MODE/winfw.dll"
-            "windows/driverlogic/bin/x64-$CPP_BUILD_MODE/driverlogic.exe"
-            # The nsis plugin is always built in 32 bit release mode
-            windows/nsis-plugins/bin/Win32-Release/*.dll
-        )
-        sign_win "${CPP_BINARIES[@]}"
-    fi
+        log_header "Building C++ code in $CPP_BUILD_MODE mode for $CPP_BUILD_TARGET"
+        CPP_BUILD_MODES=$CPP_BUILD_MODE CPP_BUILD_TARGETS=$CPP_BUILD_TARGET IS_RELEASE=$IS_RELEASE ./build-windows-modules.sh
+
+        if [[ "$SIGN" == "true" ]]; then
+            CPP_BINARIES=(
+                "windows/winfw/bin/$CPP_BUILD_TARGET-$CPP_BUILD_MODE/winfw.dll"
+                "windows/driverlogic/bin/$CPP_BUILD_TARGET-$CPP_BUILD_MODE/driverlogic.exe"
+                # The nsis plugin is always built in 32 bit release mode
+                windows/nsis-plugins/bin/Win32-Release/*.dll
+            )
+            sign_win "${CPP_BINARIES[@]}"
+        fi
+    done
 fi
 
 for t in "${TARGETS[@]:-""}"; do